Update and rename README.md to MikalaniJani
I'm a music student at Macomb Community College with plans to transfer to the University of Michigan in Fall 2027 under the Go Blue Guarantee. My goal is to earn a Bachelor and Master in Music & Technology while double majoring in Computer Science (CS-LSA). I'm currently learning full-stack and cross-platform development to build tools that empower artists and musicians. Passionate about creative coding, vocal performance, and building a portfolio that bridges art and tech.